# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Dupli1 API
  version: 1.0.0
  description: |
    Cross-service index of the Dupli1 fashion bag marketplace API, annotated with the
    permission each route requires (`x-required-permissions`).

    **Scope:** this file is a gateway-level route index, not the full contract. Request and
    response schemas live in the per-service specs, which are authoritative:
    [auth](../api/specs/auth-v1.yaml), [product](../api/specs/product-v1.yaml)
    (including inventory, catalog and coupons), [order](../api/specs/order-v1.yaml)
    (including checkout sessions) and [notification](../api/specs/notification-v1.yaml).
    Cart and payment have no dedicated spec yet — see [api.md](api.md) and
    [payment-service.md](payment-service.md).

    **Authorization:** protected routes require `Authorization: Bearer <access_token>`.
    Access tokens include a `permissions` string array. See [permissions.md](permissions.md) for the catalog.
    An empty `security` block with no `x-required-permissions` means any authenticated caller,
    subject to ABAC on their own records.

    **Paths:** routes are namespaced by owning service. Older top-level prefixes
    (`/api/v1/inventory`, `/api/v1/catalog`, `/api/v1/coupons`, `/api/v1/variants`,
    `/api/v1/checkout`, `/api/v1/carts`) remain registered as aliases but are omitted here.
    Per-service `/health` and `/settings` routes are public and also omitted, except
...

...
servers:
  - url: http://localhost:8080
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        RS256 access token from `POST /api/v1/auth/refresh`.
        Claims: `sub`, `type: access`, `permissions` (string[]).
paths:
...
  /api/v1/auth/me:
    get:
      summary: Get the authenticated user account
      tags: [Auth]
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Success
...
  /api/v1/auth/register:
    post:
      summary: Register a user
      tags: [Auth]
      security:
        - bearerAuth: []
      x-required-permissions: [user.create]
      description: Open register allows unauthenticated customer signup when AUTH_OPEN_REGISTER=true.
      responses:
        '201':
          description: Created
  /api/v1/auth/users:
    get:
      summary: List users
      tags: [Auth Admin]
      security:
        - bearerAuth: []
      x-required-permissions: [user.read]
      responses:
        '200':
          description: Success
  /api/v1/auth/users/{id}/permissions:
    parameters:
      - name: id
        in: path
        required: true
        description: Resource id
        schema:
          type: string
    patch:
      summary: Replace user permissions
      tags: [Auth Admin]
      security:
        - bearerAuth: []
      x-required-permissions: [user.permissions.update]
      responses:
        '200':
          description: Success
  /api/v1/auth/users/{id}/password:
    parameters:
      - name: id
        in: path
        required: true
        description: Resource id
        schema:
          type: string
    patch:
      summary: Set a user password
      tags: [Auth Admin]
      security:
        - bearerAuth: []
      x-required-permissions: [user.password.update]
      responses:
        '200':
          description: Success
  /api/v1/auth/users/{id}/status:
    parameters:
      - name: id
        in: path
        required: true
        description: Resource id
        schema:
          type: string
    patch:
      summary: Activate or deactivate a user
      tags: [Auth Admin]
      security:
        - bearerAuth: []
      x-required-permissions: [user.status.update]
      responses:
        '200':
          description: Success
...